BlockFileLoaderBitcoindTest: move to package `o.b.test.bitcoind`
Also add JavaDoc to new package.
